Finalizing image design
My goal is to allow this Docker image to function in two modes.  The first will
allow the complete FMI site to be built.  The second will put the image in
'server mode' where initialization is done when the image is launched, but the
actual build only happens in response to a webhook call.